There are two modal light boxes here.
The assignment was to create a lightbox/slideshow on our one_page_website. I eventually got it to work, but it doesn't look real pretty. The other is Modal_CSS, html, and JS, that was copied directly from the w3school's example. That one looks a little better, but all the styling and JS in contained within the html.